examples/requirements.txt
# Azure SDK for Python
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-monitor>=6.0.0

# Optional: For enhanced functionality
azure-cli-core>=2.50.0
tabulate>=0.9.0
colorama>=0.4.6
